+++ common/src/main/java/traben/solid_mobs/SolidMobsCrossPlatformHelper.java
@@ -1,6 +1,7 @@
 package traben.solid_mobs;
 
 import dev.architectury.injectables.annotations.ExpectPlatform;
+import net.minecraft.server.network.ServerPlayerEntity;
 
 import java.nio.file.Path;
 
@@ -10,4 +11,8 @@ public class SolidMobsCrossPlatformHelper {
     public static Path getConfigDirectory() {
         return Path.of("");
     }
+
+    @ExpectPlatform
+    public static void sendConfigToClient(@SuppressWarnings("unused") ServerPlayerEntity player){}
+
 }