+++ common/src/main/java/traben/solid_mobs/mixin/MixinEntity.java
@@ -11,11 +11,12 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+import traben.solid_mobs.SolidMobsMain;
 
 import java.util.List;
 
-import static traben.solid_mobs.solidMobsMain.EXEMPT_ENTITIES;
-import static traben.solid_mobs.solidMobsMain.solidMobsConfigData;
+import static traben.solid_mobs.SolidMobsMain.EXEMPT_ENTITIES;
+import static traben.solid_mobs.SolidMobsMain.solidMobsConfigData;
 
 @Mixin(Entity.class)
 public abstract class MixinEntity {
@@ -45,6 +46,9 @@ public abstract class MixinEntity {
     @Shadow public abstract boolean isLiving();
 
 
+    @Shadow public abstract boolean isSneaking();
+
+
     @Inject(method = "isCollidable", cancellable = true, at = @At("RETURN"))
     private void sm$collisionOverride(CallbackInfoReturnable<Boolean> cir) {
         if(solidMobsConfigData.canUseMod(this.getWorld())) {
@@ -66,15 +70,15 @@ public abstract class MixinEntity {
                     //System.out.println("true for "+getName().getString());
                     returnValue = false;
                 } else {//return false if invisible
-
-                    if(solidMobsConfigData.allowInvisibleCollisions){
+                    if(!solidMobsConfigData.allowNonSavingEntityCollisions && !getType().isSaveable() && !getType().equals(EntityType.PLAYER)){
+                        returnValue = false;
+                    }else if(solidMobsConfigData.allowInvisibleCollisions){
                         returnValue = isAlive();
                     }else{
                         returnValue = isAlive() && !isInvisible();
                     }
                     //returnValue = isAlive() && !(!solidMobsConfigData.allowInvisibleCollisions && isInvisible());
                 }
-
                 cir.setReturnValue(returnValue);
             }
         }
@@ -83,20 +87,35 @@ public abstract class MixinEntity {
     @Inject(method = "collidesWith", cancellable = true, at = @At("RETURN"))
     private void sm$collisionOverride(Entity other, CallbackInfoReturnable<Boolean> cir) {
         if (solidMobsConfigData.canUseMod(this.getWorld())) {
+            boolean collides = cir.getReturnValue();
             if (EXEMPT_ENTITIES.contains(getType().toString()) || EXEMPT_ENTITIES.contains(other.getType().toString())) {
-                cir.setReturnValue(false);
+                collides = false;
             } else if (this.isLiving() && other.isLiving()) {
                 if (getType().equals(EntityType.PLAYER) && other.getType().equals(EntityType.PLAYER)) {
                     //only affect player on player collisions we still need other things to collide with players so PLAYER cannot be in exempt list
                     if (solidMobsConfigData.allowPlayerCollisions) {
-                        cir.setReturnValue(!isConnectedThroughVehicle(other));
+                        collides = !isConnectedThroughVehicle(other);
                     } else {
-                        cir.setReturnValue(false);
+                        collides = false;
                     }
                 }
             }
+            if(solidMobsConfigData.platformMode && collides){
+                if(isSneaking()){
+                    SolidMobsMain.registerCollision(getType().toString(),other.getType().toString(),false);
+                    cir.setReturnValue(false);
+                }else {
+                    collides = getY() + 0.01 >= other.getY() + other.getBoundingBox().getYLength();
+                    SolidMobsMain.registerCollision(getType().toString(), other.getType().toString(), collides);
+                    cir.setReturnValue(collides);
+                }
+            }else {
+                SolidMobsMain.registerCollision(getType().toString(), other.getType().toString(), collides);
+                cir.setReturnValue(collides);
+            }
         }
     }
+
 
 
 
@@ -106,21 +125,10 @@ public abstract class MixinEntity {
     private void sm$moveWalkingRider(CallbackInfo ci) {
         if(solidMobsConfigData.canUseMod(this.getWorld()) && this.isLiving()) {
             if (!EXEMPT_ENTITIES.contains(getType().toString())) {
-                if ((getType().equals(EntityType.SLIME) || getType().equals(EntityType.MAGMA_CUBE)) && solidMobsConfigData.bouncySlimes) {
-//                List<Entity> colliders = getEntityWorld().getOtherEntities(((Entity) (Object) this), boundingBox.expand(-0.03, 1, -0.03));
-//                if (!colliders.isEmpty()) {
-//                    for (Entity possibleBouncingMob :
-//                            colliders) {
-//                        if (possibleBouncingMob.getY() >= this.getY() + this.getHeight()
-//                            && !possibleBouncingMob.isSneaking()
-//                        ) {
-//                            bounceUp(possibleBouncingMob);
-//                        }
-//                    }
-//                }
-                } else {//move with mob
+
+                if ((!getType().equals(EntityType.SLIME) && !getType().equals(EntityType.MAGMA_CUBE)) || !solidMobsConfigData.bouncySlimes) {//move with mob
                     try {
-                        List<Entity> colliders = getEntityWorld().getOtherEntities(((Entity) (Object) this), boundingBox.expand(-0.03, 0.1, -0.03));
+                        @SuppressWarnings("DataFlowIssue") List<Entity> colliders = getEntityWorld().getOtherEntities(((Entity) (Object) this), boundingBox.expand(-0.03, 0.1, -0.03));
                         if (!colliders.isEmpty()) {
                             //remove any vanilla passengers from this check
 //                List<Entity> vanillaRiders = getPassengerList();
@@ -135,20 +143,21 @@ public abstract class MixinEntity {
                                 ) {
                                     //apply movement to this mob
                                     if (possibleStandingMob.isSneaking()) {
-                                        //if sneaking snap player to centre of mob
-                                        // prevents falling off cause sneaking
-                                        double modifyY;
-                                        if (getType().equals(EntityType.GHAST)) {
-                                            modifyY = 0.08;
-                                        } else {
-                                            modifyY = 1.0 / 16;
+                                        if(!solidMobsConfigData.platformMode){
+                                            //if sneaking snap player to centre of mob
+                                            // prevents falling off cause sneaking
+                                            double modifyY;
+                                            if (getType().equals(EntityType.GHAST)) {
+                                                modifyY = 0.08;
+                                            } else {
+                                                modifyY = 1.0 / 16;
+                                            }
+                                            possibleStandingMob.setPosition(
+                                                    this.getPos().getX(),
+                                                    this.getPos().getY() + getHeight() + modifyY,
+                                                    this.getPos().getZ());
                                         }
-                                        possibleStandingMob.setPosition(
-                                                this.getPos().getX(),
-                                                this.getPos().getY() + getHeight() + modifyY,
-                                                this.getPos().getZ());
                                     } else {//not sneaking
-
                                         possibleStandingMob.addVelocity(
                                                 this.getVelocity().getX() * 0.833333333333,
                                                 this.getVelocity().getY() * 0.833333333333,
@@ -160,7 +169,19 @@ public abstract class MixinEntity {
                     }catch(Exception e){
                         //
                     }
-                }
+                }// else {
+//                List<Entity> colliders = getEntityWorld().getOtherEntities(((Entity) (Object) this), boundingBox.expand(-0.03, 1, -0.03));
+//                if (!colliders.isEmpty()) {
+//                    for (Entity possibleBouncingMob :
+//                            colliders) {
+//                        if (possibleBouncingMob.getY() >= this.getY() + this.getHeight()
+//                            && !possibleBouncingMob.isSneaking()
+//                        ) {
+//                            bounceUp(possibleBouncingMob);
+//                        }
+//                    }
+//                }
+//                }
             }
         }
     }