+++ common/src/main/java/traben/solid_mobs/mixin/MixinMinecraftClient.java
@@ -6,8 +6,8 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import traben.solid_mobs.SolidMobsMain;
 import traben.solid_mobs.client.solidMobsClient;
-import traben.solid_mobs.solidMobsMain;
 
 
 @Mixin(MinecraftClient.class)
@@ -17,7 +17,7 @@ public abstract class MixinMinecraftClient {
     private void etf$injected(Screen screen, CallbackInfo ci) {
         //reset to remove possible server config data
         solidMobsClient.haveServerConfig = false;
-        solidMobsMain.sm$loadConfig();
-        solidMobsMain.lastPushTime.clear();
+        SolidMobsMain.sm$loadConfig();
+        SolidMobsMain.lastPushTime.clear();
     }
 }