+++ common/src/main/java/traben/solid_mobs/mixin/MixinPlayerEntity.java
@@ -10,16 +10,15 @@ import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
 import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import static traben.solid_mobs.solidMobsMain.*;
+import static traben.solid_mobs.SolidMobsMain.*;
 
 @Mixin(PlayerEntity.class)
 public abstract class MixinPlayerEntity extends LivingEntity {
-
-
 
 
     protected MixinPlayerEntity(EntityType<? extends LivingEntity> entityType, World world) {
@@ -30,6 +29,7 @@ public abstract class MixinPlayerEntity extends LivingEntity {
 
     @Inject(method = "interact", at = @At("HEAD"))
     private void sm$tryPush(Entity entity, Hand hand, CallbackInfoReturnable<ActionResult> cir) {
+        if(entity==null || entity.isRemoved()) return;
         if (solidMobsConfigData.canUseMod(this.getWorld())
                 && entity instanceof LivingEntity alive
                 && solidMobsConfigData.allowShovingMobs
@@ -49,15 +49,18 @@ public abstract class MixinPlayerEntity extends LivingEntity {
 
 
 
+    @Unique
     private void sm$pushThisServer(LivingEntity entity){
+        if(entity == null || entity.isRemoved()) return;
         if (distanceTo(entity) < 2.5){
-
+            try {
                 if (entity instanceof VillagerEntity villager) {
                     villager.takeKnockback(0.15d, -(entity.getX() - this.getX()), -(entity.getZ() - this.getZ()));
                     lastPushTime.put(getUuid(), world.getTime() + (solidMobsConfigData.shoveAgainTimeInTicks * 2L));
                     swingHand(Hand.MAIN_HAND);
                 } else if (entity instanceof HostileEntity enemy) {
-                    enemy.damage(enemy.getDamageSources().playerAttack((PlayerEntity) (Object) this), 0);
+                    //noinspection DataFlowIssue
+                    enemy.damage(enemy.getDamageSources().playerAttack((PlayerEntity) ((Object) this)), 0);
                     //enemy.setAttacker((PlayerEntity) (Object)this);
                     //enemy.takeKnockback(0.5d, -(entity.getX() - this.getX()), -(entity.getZ() - this.getZ()));
                     lastPushTime.put(getUuid(), world.getTime() + (solidMobsConfigData.shoveAgainTimeInTicks * 5L));
@@ -67,7 +70,9 @@ public abstract class MixinPlayerEntity extends LivingEntity {
                     lastPushTime.put(getUuid(), world.getTime());
                     swingHand(Hand.MAIN_HAND);
                 }
+            }catch (Exception ignored){
 
+            }
         }
     }