+++ forge/src/main/java/traben/solid_mobs/forge/SolidMobsForge.java
@@ -1,7 +1,13 @@
 package traben.solid_mobs.forge;
 
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.RegisterCommandsEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import traben.solid_mobs.solidMobsMain;
+import net.minecraftforge.fml.loading.FMLEnvironment;
+import traben.solid_mobs.SolidMobsMain;
+import traben.solid_mobs.config.SolidMobsCommands;
 
 @Mod("solid_mobs")
 public class SolidMobsForge {
@@ -9,11 +15,21 @@ public class SolidMobsForge {
 
         ForgePacketHandler.init();
 
-        solidMobsMain.init();
-//        if (FMLEnvironment.dist == Dist.CLIENT) {
-//
-//        } else {
-//
-//        }
+        SolidMobsMain.init();
+        if (FMLEnvironment.dist != Dist.CLIENT) {
+            //server
+            //EventBuses.registerModEventBus("solid_mobs", FMLJavaModLoadingContext.get().getModEventBus());
+            MinecraftForge.EVENT_BUS.register(this);
+
+
+        }
     }
+
+    @SubscribeEvent
+    public static void onRegisterCommandEvent(RegisterCommandsEvent event){
+        SolidMobsCommands.registerCommands(event.getDispatcher(),null,null);
+    }
+
+
+
 }